backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.23
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
SQLAlchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.13
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
pgvector>=0.2.5
httpx>=0.27
email-validator>=2.2 ; python_version >= "3.8"
